pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.201.1",
    "crewai-tools>=0.75.0",
    "fastapi>=0.118.0",
    "pydantic>=2.11.10",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.2",
    "requests>=2.32.5",
    "sendgrid>=6.12.5",
    "uvicorn>=0.37.0",
]
